Hardcode list of "relevant" table names for ingestion from GitHub
The `airbyte-github` plugin by default imports 163 tables into
Splitgraph, but we only need a few of them for the analytics queries
we want to make in the demo app. So, hardcode the list of those,
but also hardcode the list of all 163 tables for reference, and also
the 43 tables that are imported given the relevant tables (because either
they depend on them via a foreign key relationship, or they're an
airbyte meta table).

For the 43 tables, see this recent import of `splitgraph/seafowl`:

* https://www.splitgraph.com/miles/github-import-splitgraph-seafowl/20230526-224723/-/tables

This took 3 minutes and 40 seconds to import into Splitgraph.