Only poll unique export taskIds, and do the polling from one hook instead of each component
Previously, the API always returned a unique taskId for each table being exported,
but a recent change optimized it to return one taskId for the set of tables being
exported, but still one taskId for each query being exported. Also previously, this
demo code rendered a loading component for each table, and each component had its
own hook for polling the taskId of that table. But now that multiple tables can share
a taskId, it doesn't make sense for each component to poll for its own taskId.

Now, we track the set of taskIds separately from the set of completed tables, and
we only poll for unique taskIds, which we do in a hook instead of in each component. And
each table preview checks the set of completed tables to know whether it's been completed.