Add workaround for failed exports of queries: create a fallback table
For each query to export, optionally provide a fallback `CREATE TABLE`
query which will run if the export job for the query fails. Implement
this by calling an API route `/api/create-fallback-table-after-failed-export`
after an export for a query fails for any reason.

This works around the bug where queries with an empty result fail
to export to Seafowl, see: https://github.com/splitgraph/seafowl/issues/423